Address comments, and add unittests for property subscribers, load balancer state and JMX classes

+++ d2/src/main/java/com/linkedin/d2/jmx/ClusterInfoJmx.java
@@ -29,7 +29,7 @@ public class ClusterInfoJmx implements ClusterInfoJmxMBean
   }
 
   @Override
-  public long getCanaryDistributionPolicy()
+  public int getCanaryDistributionPolicy()
   {
     switch (_clusterInfoItem.getClusterPropertiesItem().getDistribution())
     {
