+++ d2/src/main/java/com/linkedin/d2/jmx/ServicePropertiesJmxMBean.java
@@ -19,8 +19,8 @@ package com.linkedin.d2.jmx;
 public interface ServicePropertiesJmxMBean {
   /**
    *
-   * @return integer value of canary distribution policy when building the cluster properties:
+   * @return integer value of canary distribution policy when building the service properties:
    *         0 - STABLE, 1 - CANARY, -1 - UNSPECIFIED
    */
-  long getCanaryDistributionPolicy();
+  int getCanaryDistributionPolicy();
 }
